Big Changes & Additions
Updated adding a patient input. Added high and low threshold input for
patient. Also added the functionality for our system to be able to send
an email to the newly added patient or user with their id and password
once they are added to the system so that they can login.